Stable diffusion 3.5 medium
Runs on free colab gpu. Not the fastest but works none the less.

To use you need to generate a free hugging face token. Place that in the login. Then switch the runtime to gpu. You'll have to restart the runtime after the first cell finishes. The second cell takes around five minutes.

I'm hoping to use this as the starting point for future projects!